move: 도메인별로 디렉토리 분리
